Move file to sceanrio folder
